fix: model filenames
